pyproject.toml
[project]
name = "pmkit"
version = "0.1.0"
description = "Product Management Toolkit wrapping OpenProject"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
pmkit = "pmkit.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
